pyproject.toml
[project]
name = "mas-rd-course-9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"

requires-python = ">=3.12"

dependencies = [
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "faiss-cpu>=1.8.0",
    "rank-bm25>=0.2.2",
    "python-dotenv>=1.0.1",
    "trafilatura>=1.12.0",
    "duckduckgo-search>=6.0.0",
    "sentence-transformers>=3.0.0",
    "pypdf>=4.3.0",
    "ddgs>=9.12.0",
    "fastmcp>=3.2.4",
    "acp-sdk>=1.0.3",
]
